tests/requirements/latest.txt
django
sorl-thumbnail
Pillow